Added server shrinkage option
The SHRINK_SERVER flags determines whether the new server to retain the previous neurons or to create a new server with new weights. In practice it doesn't seem to have any effect. It is trivial so it kinda makes sense that the shrinkage doesn't do anything.

Also added seed for reproducability.