Updated VFL server to adjust automatically size
The VFL server adjusts the architecture according to the number of inputs it receives.
The assumption is that when it receives fewer inputs it means that one client is removed from the process but the process should continue.
Similarly when one client is added the model can be adjusted.

The change in architecture affects negatively the accuracy/r2 score but for the demo is fine.

Can reconsider more clever mechanisms to change the architecture (mainly savepoints, a better shrinking method)

Mechanisms tested locally. Online I need to reexamine the disagreement resolution (aka, if it continues with less clients).