Added policy reintroduction from flag
Reintroduces the policy that was removed from the policy removal (renamed from change_policies).

Had to specify the correct dockerhub images in the yaml helm charts.

Tested the removal of the policy on round 20 and then the reintroduction on round 30 and it works fine. However, now the model is untrained (probably it is automatically resizing to 2 clients and then re-adds the client back in).