Dynamic add/subtraction of approved clients
Api gateway propagates the training request to the approved clients.
This is a design choice that says that even if not all the clients (from the original request) are approved the process can still continue. Of course this requires that the aggregator (server) can handle this.

The code can handle both the removal and the addition of the clients according to the message from the policy enforcer.