Changed the api-gateway to a polling mechanism
the original request returns the request ID.
Then a new endpoint was introduced that enables polling. The polling result is a list of accuracies, nof_clients, timestamps, and metadata. The polling is updated after every round.

Also there is a flag that determines whether the process is done or not. If it is not done the status is pending and no new request is allowed. (Only one active job at any time).

Note:
- it takes some time to update the status to done, when it finished. (1 minute I believe, probably because of a timeout/ time buffer)
- the metadata also are updated towards the end, I should change that so that I have them available earlier.
- right now there might be some residual responses not fully utilized from the previous approach
- only tested on VFL, not on the other use cases
- error handling implemented but might need some extra testing