updated scripts from jupyter and live plotting
Saved the management tools and other scripts from the jupyter.

Added plotting functionality to be paired with the polling mechanism. The mechanism polls the results every 2 seconds and plots the r2 score. The plots are not that pretty but work OK.

Could consider smoothing to make the results a little better. Also I could rework the training parameters for better results maybe.